+++ program-runtime/src/lib.rs
@@ -1,6 +1,7 @@
 #![cfg_attr(RUSTC_WITH_SPECIALIZATION, feature(min_specialization))]
 
 pub mod accounts_data_meter;
+pub mod compute_budget;
 pub mod instruction_recorder;
 pub mod invoke_context;
 pub mod log_collector;
